Fixes erroneous LaTeX code generation for verse.
If new bible verse began at the start of a paragraph which was also in
verse environment, the \begin{verse} statement came proor to the
\switchcolumn command, resulting in LaTeX errors and failure to
compile.

Slight bodge of a fix. Rather than using a more complex regexp to
account for this (initial attempt didn't succeed), scrub text if
necessary from end of one paragraph and stick it onto the next. Not very
neat but works for now.